fix bug

+++ cinn/hlir/op/contrib/argmax_test.cc
@@ -65,7 +65,7 @@ TEST(GenerateCode_Cpu, Argmax) {
   ir::Expr w(28);
 
   lang::Placeholder<float> in("in", {n, in_c, h, w});
-  lang::Placeholder<float> out("out", {n, h, w});
+  lang::Placeholder<int> out("out", {n, h, w});
   ir::Tensor res = Argmax(in, axis, true, "test_argmax_in");
 
   poly::StageMap stages = poly::CreateStages({res});
